pyproject.toml
[project]
name = "nano-openclaw"
version = "0.1.0"
description = "Minimal Python reimplementation of OpenClaw's agent loop, for learning."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "caft", email = "caft0505@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40",
    "openai>=1.0",
    "rich>=13.7",
    "Pillow>=10.0",
    "json5>=0.9",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "prompt_toolkit>=3.0",
    "mcp>=1.0.0",
    "httpx>=0.27",
    "ddgs>=9.0",
    "readability-lxml>=0.8",
]

[project.scripts]
nano-openclaw = "nano_openclaw.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nano_openclaw"]
